backend/requirements.txt
# Flask Core
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0

# Database
pymysql==1.1.0
cryptography==41.0.7

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Vector Database
chromadb==0.4.22

# LangChain
langchain==0.1.0
langchain-community==0.0.10

# File Processing
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1

# HTTP & Utils
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1

# RAG Enhancement
rank-bm25==0.2.2
jieba==0.42.1
numpy>=1.24.0

# PDF Export
reportlab==4.0.8

# Async
gevent==23.9.1
